lesson_11/app/requirements.txt
Flask==2.1.0
gunicorn==20.1.0
Pillow==9.2.0
